chore: set up structure for publication
- remove graphviz-react library, just have one library (graphviz) that exports everything
- remove jest dep from root
- add jest/ts-jest to graphviz library
- add aliases to root package.json to run commands from root
