--- packages/graphviz-react/README.md
@@ -1,1 +1,0 @@
-# graphviz-react
