--- packages/graphviz-react/package.json
@@ -1,25 +1,0 @@
-{
-	"name": "@thingco/graphviz-react",
-	"description": "React renderer bindings for the @thingco/graphviz library",
-	"version": "0.0.0",
-	"main": "lib/index.js",
-	"module": "lib/index.module.js",
-	"types": "lib/index.d.ts",
-	"public": true,
-	"scripts": {
-		"start": "yarn rimraf lib && yarn tsc --watch",
-		"build": "yarn rimraf lib && yarn tsc"
-	},
-	"devDependencies": {
-		"@thingco/graphviz": "workspace:packages/graphviz",
-		"@types/react": "^17.0.3",
-		"@types/rimraf": "^3.0.0",
-		"react": "^17.0.1",
-		"rimraf": "^3.0.2",
-		"typescript": "^4.2.3"
-	},
-	"peerDependencies": {
-		"@thingco/graphviz": "*",
-		"react": "*"
-	}
-}
