--- packages/graphviz-react/tsconfig.json
@@ -1,9 +1,0 @@
-{
-	"extends": "../../tsconfig.json",
-	"compilerOptions": {
-		"outDir": "./lib",
-		"rootDir": "./src"
-	},
-	"include": ["src/**/*"],
-	"exclude": ["src/**/*.test.{ts,tsx}"]
-}
