+++ packages/graphviz/package.json
@@ -1,17 +1,32 @@
 {
 	"name": "@thingco/graphviz",
-	"description": "A small data visualisation library for ThingCo front ends. NOT a renderer, this just takes data and processes it to a form usable by a renderer",
+	"description": "React-based graph/chart visualisations",
 	"version": "0.0.0",
 	"main": "lib/index.js",
+	"module": "lib/index.js",
 	"types": "lib/index.d.ts",
-	"public": true,
+	"public": false,
+	"publishConfig": {
+		"registry": "https://npm.pkg.github.com/"
+	},
+	"repository": "https://github.com/thingco/shared-frontend-libs/tree/master/packages/graphviz",
 	"scripts": {
 		"start": "yarn rimraf lib && yarn tsc --watch",
-		"build": "yarn rimraf lib && yarn tsc"
+		"build": "yarn rimraf lib && yarn tsc",
+		"test": "yarn jest"
 	},
 	"devDependencies": {
+		"@thingco/graphviz": "workspace:packages/graphviz",
+		"@types/react": "^17.0.3",
 		"@types/rimraf": "^3.0.0",
+		"jest": "^26.6.3",
+		"react": "^17.0.1",
 		"rimraf": "^3.0.2",
+		"ts-jest": "^26.5.3",
 		"typescript": "^4.2.3"
+	},
+	"peerDependencies": {
+		"@thingco/graphviz": "*",
+		"react": "*"
 	}
 }
