+++ packages/graphviz/src/core/create-graph.ts
@@ -82,9 +82,7 @@ export function createGraph({
 	 * If neither the x and y axis values are provided, cannot create the graph:
 	 */
 	if (!xAxisValues && !yAxisValues) {
-		throw new Error(
-			"ONE OR BOTH x- or y-axis value arrays MUST be provided to create the graph."
-		);
+		throw new Error("ONE OR BOTH x- or y-axis value arrays MUST be provided to create the graph.");
 	}
 
 	/**
@@ -92,15 +90,9 @@ export function createGraph({
 	 * to be the length of the other (_ie_ the values are plotted on a range/in series):
 	 */
 	if (yAxisValues && !xAxisValues) {
-		xAxisValues = Array.from(
-			{ length: yAxisValues.length },
-			(_, i) => i + xAxisMin
-		);
+		xAxisValues = Array.from({ length: yAxisValues.length }, (_, i) => i + xAxisMin);
 	} else if (xAxisValues && !yAxisValues) {
-		yAxisValues = Array.from(
-			{ length: xAxisValues.length },
-			(_, i) => i + yAxisMin
-		);
+		yAxisValues = Array.from({ length: xAxisValues.length }, (_, i) => i + yAxisMin);
 	}
 	// Once both of the arrays of values are populated, can infer the max values if not present:
 	xAxisMax = xAxisMax ?? Math.max(...(xAxisValues as number[]));
@@ -176,11 +168,7 @@ export function createGraph({
  * @param {number} axisMax - the value the axis ends on
  * @returns {number}
  */
-export function calculateScale(
-	axisSize: number,
-	axisMin: number,
-	axisMax: number
-): number {
+export function calculateScale(axisSize: number, axisMin: number, axisMax: number): number {
 	return axisSize / (axisMax - axisMin);
 }
 
@@ -206,11 +194,7 @@ export function calculateScale(
  * @param {number} axisMax - the value the axis ends on
  * @returns {number}
  */
-export function calculateSize(
-	axisScale: number,
-	axisMin: number,
-	axisMax: number
-): number {
+export function calculateSize(axisScale: number, axisMin: number, axisMax: number): number {
 	return (axisMax - axisMin) * axisScale;
 }
 
@@ -223,11 +207,7 @@ export function calculateSize(
  * @param {number} graphData.xAxisStep
  * @returns {number[]}
  */
-export function steppedXAxisValues({
-	xAxisMin,
-	xAxisMax,
-	xAxisStep,
-}: GraphData): number[] {
+export function steppedXAxisValues({ xAxisMin, xAxisMax, xAxisStep }: GraphData): number[] {
 	const vals = [];
 
 	// TODO currently ignoring graphs that only have data with negative values that should be plotted
@@ -252,11 +232,7 @@ export function steppedXAxisValues({
  * @param {number} graphData.yAxisStep
  * @returns {number[]}
  */
-export function steppedYAxisValues({
-	yAxisMin,
-	yAxisMax,
-	yAxisStep,
-}: GraphData): number[] {
+export function steppedYAxisValues({ yAxisMin, yAxisMax, yAxisStep }: GraphData): number[] {
 	const vals = [];
 
 	// TODO currently ignoring graphs that only have data with negative values that should be plotted
