+++ packages/graphviz/src/core/index.ts
@@ -1,0 +1,2 @@
+export * from "./create-graph";
+export * from "./projections/svg";
