+++ packages/graphviz/src/react/Annotations.tsx
@@ -1,12 +1,12 @@
-import {
-  steppedXAxisValues,
-  steppedYAxisValues,
-  VerticalAlignment,
-  xAxisAnnotations,
-  yAxisAnnotations
-} from "@thingco/graphviz";
 import React from "react";
 
+import {
+	steppedXAxisValues,
+	steppedYAxisValues,
+	VerticalAlignment,
+	xAxisAnnotations,
+	yAxisAnnotations,
+} from "../core";
 import { useGraph } from "./Context";
 
 export interface AxisAnnotationsProps {
