+++ packages/graphviz/src/react/Axes.tsx
@@ -1,6 +1,6 @@
-import { VerticalAlignment, xAxis, xAxisSteps, yAxis, yAxisSteps } from "@thingco/graphviz";
 import React from "react";
 
+import { VerticalAlignment, xAxis, xAxisSteps, yAxis, yAxisSteps } from "../core";
 import { useGraph } from "./Context";
 
 export interface AxisProps {
@@ -28,23 +28,13 @@ export const XAxis = ({
 		<g data-componentid="x-axis">
 			<line style={style} x1={x1} x2={x2} y1={y1} y2={y2} />
 			{axisSteps?.map(({ x1, x2, y1, y2 }, i) => (
-				<line
-					key={`${x1}${x2}${y1}${y2}${i}`}
-					style={style}
-					x1={x1}
-					x2={x2}
-					y1={y1}
-					y2={y2}
-				/>
+				<line key={`${x1}${x2}${y1}${y2}${i}`} style={style} x1={x1} x2={x2} y1={y1} y2={y2} />
 			))}
 		</g>
 	);
 };
 
-export const YAxis = ({
-	style = defaultAxisStyle,
-	showSteps = true,
-}: AxisProps): JSX.Element => {
+export const YAxis = ({ style = defaultAxisStyle, showSteps = true }: AxisProps): JSX.Element => {
 	const graph = useGraph();
 	const { x1, x2, y1, y2 } = yAxis(graph);
 	const axisSteps = showSteps ? yAxisSteps(graph) : null;
@@ -53,14 +43,7 @@ export const YAxis = ({
 		<g data-componentid="y-axis">
 			<line style={style} x1={x1} x2={x2} y1={y1} y2={y2} />
 			{axisSteps?.map(({ x1, x2, y1, y2 }, i) => (
-				<line
-					key={`${x1}${x2}${y1}${y2}${i}`}
-					style={style}
-					x1={x1}
-					x2={x2}
-					y1={y1}
-					y2={y2}
-				/>
+				<line key={`${x1}${x2}${y1}${y2}${i}`} style={style} x1={x1} x2={x2} y1={y1} y2={y2} />
 			))}
 		</g>
 	);
