+++ packages/graphviz/src/react/Context.tsx
@@ -1,5 +1,6 @@
-import { GraphData } from "@thingco/graphviz";
 import React from "react";
+
+import { GraphData } from "../core";
 
 /**
  * NOTE the context is defined apart from the Provider to prevent circular dependencies
@@ -7,13 +8,14 @@ import React from "react";
  */
 export const GraphContext = React.createContext<GraphData | null>(null);
 
+/**
+ *
+ */
 export function useGraph(): GraphData {
 	const graph = React.useContext(GraphContext);
 
 	if (!GraphContext) {
-		throw new Error(
-			"Graph values can only be accessed from within a graph context provider."
-		);
+		throw new Error("Graph values can only be accessed from within a graph context provider.");
 	} else if (!graph) {
 		throw new Error(
 			"Graph provider value is null: cannot access anything from it, please configure it correctly."
