+++ packages/graphviz/tsconfig.json
@@ -1,11 +1,9 @@
 {
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "rootDir": "./src",
-  },
-  "include": ["src/**/*"],
-  "exclude": [
-    "src/**/*.test.{ts,tsx}"
-  ]
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "./lib",
+		"rootDir": "./src"
+	},
+	"include": ["src/**/*"],
+	"exclude": ["src/**/*.test.{ts,tsx}"]
 }
