+++ tsconfig.jest.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"module": "commonjs",
+		"target": "es2020",
+		"noEmit": true
+	}
+}
